requirements.txt
apturl==0.5.2
autoenv==1.0.0
beautifulsoup4==4.4.1
bleach==2.1.3
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
decorator==4.0.11
defer==1.0.6
entrypoints==0.2.3
feedparser==5.1.3
guacamole==0.9.2
html5lib==1.0.1
httplib2==0.9.1
idna==2.0
imageio==2.1.2
ipykernel==4.8.2
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.1.2
Jinja2==2.8
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
matplotlib==2.2.0
mistune==0.8.3
moviepy==0.2.3.2
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.1
oauthlib==1.0.3
onboard==1.2.0
opencv-python==3.4.0.12
padme==1.1.1
pandocfilters==1.4.2
pexpect==4.0.1
pickleshare==0.7.4
Pillow==5.0.0
plainbox==0.25
prompt-toolkit==1.0.15
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.6.1
python-debian==0.1.27
python-systemd==231
pytz==2018.3
pyxdg==0.25
pyzmq==17.0.0
qtconsole==4.3.1
reportlab==3.3.0
requests==2.9.1
scipy==1.0.0
Send2Trash==1.5.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
six==1.10.0
system-service==0.3
terminado==0.8.1
testpath==0.3.1
tornado==5.0
tqdm==4.11.2
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.13.1
usb-creator==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
